Modified workflow
Turns out aCropalypse is not an easy CVE to replicate. Not only am I
unable to find a quick way to replicate the vuln, Twitter also has
mitigation feature against it (by auto truncating uploaded images.

The workflow will now require the hash to be explicitly exposed, but
without the repo name. This makes the chal slightly easier and less
guessy since player no longer need to use aCropalypse, but still
requires the knowledge that commits can be accessed throughout the fork
network.

Also added some new posts to be posted on twitter.